finish in 59 minutes

+++ 7kyu/rotate_for_a_max/rotate_for_a_max_test.go
@@ -7,7 +7,7 @@ import "testing"
 func TestMaxRot(t *testing.T) {
 	cases := []struct {
 		Name        string
-		Input, Want int
+		Input, Want int64
 	}{
 		{
 			Name:  "it should return 68957",
